# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
components:
  securitySchemes:
    clerkSession:
      type: apiKey
      in: cookie
      name: __session
  schemas:
...
    OperationalReport:
      type: object
      required: [id, assignmentId, busNumber, officialRunKey, category, location, locationObservedAt, createdAt, resolvedAt, resolution]
      properties:
        id: { type: string, format: uuid }
        assignmentId: { type: integer }
        busNumber: { type: string }
        officialRunKey: { type: string }
        category: { type: string, enum: [stop_blocked, heavy_traffic, coach_issue, passenger_assistance] }
        location:
          type: object
          required: [lat, lng]
          properties:
            lat: { type: number }
            lng: { type: number }
        locationObservedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
        resolvedAt: { type: [string, 'null'], format: date-time }
        resolution: { type: [string, 'null'] }
    DriverReportInput:
      type: object
      additionalProperties: false
      required: [category]
      properties:
        category: { type: string, enum: [stop_blocked, heavy_traffic, coach_issue, passenger_assistance] }
    OperationalReportResolution:
      type: object
      additionalProperties: false
      required: [resolution]
      properties:
        resolution: { type: string, minLength: 1, maxLength: 240 }
    DispatchInstruction:
      type: object
      required: [id, assignmentId, busNumber, officialRunKey, message, createdAt, deliveredAt, acknowledgedAt]
      properties:
        id: { type: string, format: uuid }
        assignmentId: { type: integer }
        busNumber: { type: string }
        officialRunKey: { type: string }
        message: { type: string }
        createdAt: { type: string, format: date-time }
        deliveredAt: { type: [string, 'null'], format: date-time }
        acknowledgedAt: { type: [string, 'null'], format: date-time }
    DispatchInstructionInput:
      type: object
      additionalProperties: false
      required: [assignmentId, message]
      properties:
        assignmentId: { type: integer, minimum: 1 }
        message: { type: string, minLength: 1, maxLength: 500 }
    DriverCommunications:
      type: object
      required: [reports, instructions, safety]
      properties:
        reports:
          type: array
          items: { $ref: '#/components/schemas/OperationalReport' }
        instructions:
          type: array
          items: { $ref: '#/components/schemas/DispatchInstruction' }
        safety:
          type: object
          required: [stopped, reason]
          properties:
            stopped: { type: boolean }
            reason: { type: string }
    AdminCommunications:
      type: object
      required: [reports, instructions]
      properties:
        reports:
          type: array
          items: { $ref: '#/components/schemas/OperationalReport' }
        instructions:
          type: array
          items: { $ref: '#/components/schemas/DispatchInstruction' }
...
paths:
...
  /driver/communications:
    get:
      operationId: getDriverCommunications
      tags: [driver]
      security: [{ clerkSession: [] }]
      responses:
        '200':
          description: Reports and dispatch instructions scoped to the signed-in driver's current assignment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DriverCommunications' }
        '403': { description: Driver access or current assignment required }
  /driver/operational-reports:
    post:
      operationId: createDriverOperationalReport
      tags: [driver]
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DriverReportInput' }
      responses:
        '201':
          description: Server-attributed operational report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalReport' }
        '409': { description: Coach is not confirmed safely stopped }
...
  /driver/dispatch-instructions/{id}/acknowledge:
    post:
      operationId: acknowledgeDispatchInstruction
      tags: [driver]
      security: [{ clerkSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Explicit stopped-driver acknowledgment recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DispatchInstruction' }
        '409': { description: Coach is not confirmed safely stopped }
  /admin/communications:
    get:
      operationId: getAdminCommunications
      tags: [admin]
      security: [{ clerkSession: [] }]
      responses:
        '200':
          description: Operational reports and dispatch instruction delivery state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminCommunications' }
  /admin/dispatch-instructions:
    post:
      operationId: createDispatchInstruction
      tags: [admin]
      security: [{ clerkSession: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DispatchInstructionInput' }
      responses:
        '201':
          description: Persisted instruction scoped to an active assignment
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DispatchInstruction' }
        '409': { description: Assignment is no longer active }
  /admin/operational-reports/{id}/resolve:
    post:
      operationId: resolveOperationalReport
      tags: [admin]
      security: [{ clerkSession: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/OperationalReportResolution' }
      responses:
        '200':
          description: Resolved report
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OperationalReport' }
...